Consistently/correctly handle numpy integral types
Fixes: https://github.com/pytorch/pytorch/issues/29252

Summary:

There is some inconsistent/incorrect code in the handling of
parsing python/numpy scalars that led to issue #29252:

Where we allow interpretation of a zero-dim numpy integer value
as a scalar in pytorch, we incorrectly parsed the int as a float.

Also fixes the issue described here:
https://github.com/pytorch/pytorch/pull/27628/files#diff-f539198dd366265fb8dc2d661bc5d5bcR1487

And also allows numpy integral types to be used in cases where we allow
python ints (treated as long internally).

Test Plan:

Added a unit test based on the example given in the issue, a test for numpy
ints being used as ints in operator arguments, and a couple of miscelaneous
test cases.

ghstack-source-id: b34b0b6a3e17c3067b4030d15887f8e049db6980
Pull Request resolved: https://github.com/pytorch/pytorch/pull/30486