Remove BUILD_NAMEDTENSOR from codegen and .cu files
Changelist:
- remove BUILD_NAMEDTENSOR from .cu files
- remove BUILD_NAMEDTENSOR special handling in function_wrapper.py,
gen.py
- remove BUILD_NAMEDTENSOR from cpp_extension.py. This code actually
did nothing because we always compile with BUILD_NAMEDTENSOR.

Future work:
- Remove torch._C.BUILD_NAMEDTENSOR. This is used in some python
functions.
- Remove aten/src/ATen/env.py, which defines BUILD_NAMEDTENSOR=True for
codegen. There should be no use cases of this left after this PR, but it
is future work to check that.
- Remove `aten/src/ATen/core/EnableNamedTensor.h`, which defines
the BUILD_NAMEDTENSOR C++ macro. There should be no uses of this left
after this PR, but it is future work to check that

Test Plan:
- run tests

ghstack-source-id: 12136d18c7ff0279f1d80586364ef12b6711ed86
Pull Request resolved: https://github.com/pytorch/pytorch/pull/31047